Loading material from gltf
At the moment we render only the first material and the first mesh

+++ vulkan_ray_tracing/src/bottom_level_acceleration_structure.rs
@@ -83,7 +83,7 @@ impl BottomLevelAccelerationStructureBuilder {
             .index_data(self.index_buffer.unwrap())
             .index_offset(self.index_offset)
             .index_count(self.index_count)
-            .index_type(vk::IndexType::UINT16)
+            .index_type(vk::IndexType::UINT32)
             .transform_data(vk::Buffer::null())
             .transform_offset(0)
             .build();
